apps/monitoring/pod_monitor/requirements.txt
kubernetes>=28.1.0
pyvmomi>=8.0.0
prometheus-client>=0.17.0
pyyaml>=6.0
typer>=0.9.0
rich>=13.6.0
requests>=2.31.0
